refactor: restructure and clean up lib directory, unify file naming conventions
