+++ frontend/src/components/__dev/NearbyTrees.svelte
