+++ frontend/src/components/trees/index.ts
@@ -1,4 +1,3 @@
 export { default as AdoptTreeButton } from './AdoptTreeButton.svelte';
-export { default as TreeDetail } from './TreeDetail.svelte';
 export { default as TreeMetric } from './TreeMetric.svelte';
 export { default as WaterColumn } from './WaterColumn.svelte';
