+++ frontend/src/lib/geo/index.ts
@@ -1,4 +1,4 @@
 export * from './boundingBox';
 export * from './distance';
 export * from './location';
-export * from './extractTreeCandidates';
+export * from './segments';
