+++ frontend/src/lib/trees/extractNearbyTrees.ts
