+++ frontend/src/lib/trees/index.ts
@@ -1,0 +1,3 @@
+export * from './extractNearbyTrees';
+export * from './loadAdoptedTrees';
+export * from './filters';
